Cross-validation in Multiply Imputed datasets

cv_MI Cross-validation by applying multiple single imputation runs in train and test folds. Called by function psfmi_perform.

Usage

cv_MI(pobj, data_orig, folds, nimp_cv, BW, p.crit, miceImp, ...)

Arguments

pobj

An object of class pmods (pooled models), produced by a previous call to psfmi_lr.

data_orig

dataframe of original dataset that contains missing data.

folds

The number of folds, default is 3.

nimp_cv

Numerical scalar. Number of (multiple) imputation runs.

BW

If TRUE backward selection is conducted within cross-validation. Default is FALSE.

p.crit

A numerical scalar. P-value selection criterium used for backward during cross-validation. When set at 1, pooling and internal validation is done without backward selection.

miceImp

Wrapper function around the mice function.

...

Arguments as predictorMatrix, seed, maxit, etc that can be adjusted for the mice function.

See also

Author

Martijn Heymans, 2020