Function for forward selection of Linear and Logistic regression models.

glm_fw Forward selection of Linear and Logistic regression models in single dataset using as selection method the likelihood-ratio test statistic.

glm_fw(
  data,
  formula = NULL,
  Outcome = NULL,
  predictors = NULL,
  p.crit = 1,
  cat.predictors = NULL,
  spline.predictors = NULL,
  int.predictors = NULL,
  keep.predictors = NULL,
  nknots = NULL,
  model_type = "binomial"
)

Arguments

data

A data frame.

formula

A formula object to specify the model as normally used by glm. See under "Details" and "Examples" how these can be specified.

Outcome

Character vector containing the name of the outcome variable.

predictors

Character vector with the names of the predictor variables. At least one predictor variable has to be defined. Give predictors unique names and do not use predictor name combinations with numbers as, age2, gnder10, etc.

p.crit

A numerical scalar. P-value selection criterium. A value of 1 provides the full model without selection.

cat.predictors

A single string or a vector of strings to define the categorical variables. Default is NULL categorical predictors.

spline.predictors

A single string or a vector of strings to define the (restricted cubic) spline variables. Default is NULL spline predictors. See details.

int.predictors

A single string or a vector of strings with the names of the variables that form an interaction pair, separated by a “:” symbol.

keep.predictors

A single string or a vector of strings including the variables that are forced in the model during predictor selection. All type of variables are allowed.

nknots

A numerical vector that defines the number of knots for each spline predictor separately.

model_type

A character vector. If "binomial" a logistic regression model is used (default) and for "linear" a linear regression model is used.

Value

An object of class smods (single models) from which the following objects can be extracted: original dataset as data, model at each selection step RR_model, final selected model as RR_model_final, p-values at final step multiparm_final, and at each step as multiparm, formula object at final step as formula_final, and at each step as formula_step and for start model as formula_initial, predictors included at each selection step as predictors_in, predictors excluded at each step as predictors_out, and Outcome, p.crit, call,

model_type, predictors_final for names of predictors in final selection step and

predictors_initial for names of predictors in start model and keep.predictors for variables that are forced in the model during selection.

Details

A typical formula object has the form Outcome ~ terms. Categorical variables has to be defined as Outcome ~ factor(variable), restricted cubic spline variables as Outcome ~ rcs(variable, 3). Interaction terms can be defined as Outcome ~ variable1*variable2 or Outcome ~ variable1 + variable2 + variable1:variable2. All variables in the terms part have to be separated by a "+".

References

http://missingdatasolutions.rbind.io/

See also

Author

Martijn Heymans, 2021

Examples

data1 <- subset(psfmi::lbpmilr, Impnr==1) # extract first imputed dataset
res_single <- glm_fw(data=data1, p.crit = 0.05, formula=Chronic ~
       Tampascale + Smoking + factor(Satisfaction), model_type="binomial")
#> Entered at Step 1 is - Tampascale
#> Entered at Step 2 is - factor(Satisfaction)
#> 
#> Selection correctly terminated, 
#> No new variables entered the model
         
res_single$RR_model_final
#> $`Final model`
#>                          Estimate  Std Error    Z value      P-value
#> (Intercept)           -12.3880525 2.00751683 -6.1708337 6.793084e-10
#> factor(Satisfaction)2  -0.4487162 0.50581504 -0.8871153 3.750169e-01
#> factor(Satisfaction)3  -1.7678975 0.61688604 -2.8658414 4.159026e-03
#> Tampascale              0.3256887 0.05211887  6.2489583 4.131991e-10
#>                                 OR low EXP(OR) High EXP(OR)
#> (Intercept)           4.168091e-06   0.0195510    51.148704
#> factor(Satisfaction)2 6.384472e-01   0.7026122     5.103082
#> factor(Satisfaction)3 1.706915e-01   0.3540188     3.974060
#> Tampascale            1.384984e+00   3.6068366     4.424411
#> 

res_single <- glm_fw(data=data1, p.crit = 0.05, formula=Pain ~
         Tampascale  + Smoking + factor(Satisfaction), model_type="linear")
#> Entered at Step 1 is - Tampascale
#> 
#> Selection correctly terminated, 
#> No new variables entered the model
         
res_single$RR_model_final
#> $`Final model`
#>               Estimate Std. Error   t value     Pr(>|t|)
#> (Intercept) -7.2051940  0.6174338 -11.66958 4.730331e-23
#> Tampascale   0.3166061  0.0157284  20.12958 2.426905e-45
#>