Calculates the Hosmer and Lemeshow goodness of fit test.

hoslem_test the Hosmer and Lemeshow goodness of fit test.

hoslem_test(y, yhat, g = 10)

Arguments

y

a vector of observations (0/1).

yhat

a vector of predicted probabilities.

g

Number of groups tested. Default is 10. Can not be < 3.

Value

The Chi-squared test statistic, the p-value, the observed and expected frequencies.

References

Kleinman K and Horton NJ. (2014). SAS and R: Data Management, Statistical Analysis, and Graphics. 2nd Edition. Chapman & Hall/CRC.

See also

Author

Martijn Heymans, 2021

Examples

  fit <- glm(Mortality ~ Dementia + factor(Mobility) + ASA + 
   Gender + Age, data=hipstudy, family=binomial) 
   pred <- predict(fit, type = "response")
  
  hoslem_test(fit$y, pred)
#> $chisq
#> [1] 10.6652
#> 
#> $p.value
#> [1] 0.2214007
#> 
#> $observed
#>                   
#> cutyhat            V1  y
#>   [0.00182,0.0307] 44  0
#>   (0.0307,0.0562]  42  4
#>   (0.0562,0.0774]  38  0
#>   (0.0774,0.118]   41  3
#>   (0.118,0.162]    37  4
#>   (0.162,0.21]     33 12
#>   (0.21,0.287]     27 13
#>   (0.287,0.377]    27 16
#>   (0.377,0.499]    25 17
#>   (0.499,0.827]    19 24
#> 
#> $expected
#>                   
#> cutyhat                    V1       yhat
#>   [0.00182,0.0307] 43.2284119  0.7715881
#>   (0.0307,0.0562]  44.0087643  1.9912357
#>   (0.0562,0.0774]  35.4546570  2.5453430
#>   (0.0774,0.118]   39.6638651  4.3361349
#>   (0.118,0.162]    35.2967609  5.7032391
#>   (0.162,0.21]     36.6852370  8.3147630
#>   (0.21,0.287]     29.8389686 10.1610314
#>   (0.287,0.377]    28.4271628 14.5728372
#>   (0.377,0.499]    23.8043695 18.1956305
#>   (0.499,0.827]    16.5918027 26.4081973
#>