Provides adjusted intercept after shrinkage pooled coefficients in multiply imputed datasets

pool_intadj Provides pooled adjusted intercept after shrinkage pooled coefficients in multiply imputed datasets for models selected with the psfmi_lr function and internally validated with the psfmi_perform function.

pool_intadj(pobj, shrinkage_factor)

Arguments

pobj
An object of class smodsmi (selected models in multiply imputed datasets), produced by a previous call to psfmi_lr.

shrinkage_factor
A numerical scalar. Shrinkage factor value as a result of internal validation with the psfmi_perform function.

Value

A pool_intadj object from which the following objects can be extracted: int_adj, the adjusted intercept value, coef_shrink_pooled, the pooled regression coefficients after shrinkage, coef_orig_pooled, the (original) pooled regression coefficients before shrinkage and nimp, the number of imputed datasets.

Details

The function provides the pooled adjusted intercept after shrinkage of pooled regression coefficients in multiply imputed datasets. The function is only available for logistic regression models without random effects.

References

F. Harrell. Regression Modeling Strategies. With Applications to Linear Models, Logistic and Ordinal Regression, and Survival Analysis (2nd edition). Springer, New York, NY, 2015.

EW. Steyerberg (2019). Clinical Prediction MOdels. A Practical Approach to Development, Validation, and Updating (2nd edition). Springer Nature Switzerland AG.

http://missingdatasolutions.rbind.io/

Examples

 res_psfmi <- psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
           predictors=c("Gender", "Pain","Tampascale","Smoking","Function",
           "Radiation", "Age"), p.crit = 1, method="D1")

#> Registered S3 methods overwritten by 'car':
#>   method                          from
#>   influence.merMod                lme4
#>   cooks.distance.influence.merMod lme4
#>   dfbeta.influence.merMod         lme4
#>   dfbetas.influence.merMod        lme4

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

 res_psfmi$RR_Model

#> $`Step 1`
#>                 est std.err signif    lower  upper     OR   L.OR   U.OR
#> (Intercept) -5.7989  3.0757 0.0630 -11.9196 0.3218 0.0030 0.0000 1.3797
#> Gender      -0.2672  0.5041 0.5962  -1.2554 0.7211 0.7656 0.2850 2.0567
#> Pain         0.5250  0.2396 0.0325   0.0453 1.0046 1.6904 1.0464 2.7308
#> Tampascale   0.1053  0.0905 0.2485  -0.0751 0.2856 1.1110 0.9277 1.3306
#> Smoking      0.1677  0.4236 0.6923  -0.6629 0.9983 1.1826 0.5153 2.7136
#> Function    -0.0661  0.0523 0.2062  -0.1686 0.0364 0.9360 0.8448 1.0371
#> Radiation    0.2647  0.4685 0.5725  -0.6572 1.1866 1.3030 0.5183 3.2758
#> Age         -0.0158  0.0235 0.5015  -0.0618 0.0303 0.9843 0.9400 1.0307
#> 

 set.seed(100)
 res_val <- psfmi_perform(res_psfmi, method = "MI_boot", nboot=50,
   int_val = TRUE, p.crit=1, cal.plot=FALSE, plot.indiv=FALSE)

#> 
#> Calibration plot not made, only when int.val = FALSE

#> 
#> Boot 1

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 2

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 3

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 4

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 5

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 6

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 7

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 8

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 9

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 10

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 11

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 12

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 13

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 14

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 15

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 16

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 17

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 18

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 19

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 20

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 21

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 22

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 23

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 24

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 25

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 26

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 27

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 28

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 29

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 30

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 31

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 32

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 33

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 34

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 35

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 36

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 37

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 38

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 39

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 40

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 41

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 42

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 43

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 44

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 45

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 46

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 47

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 48

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 49

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

#> 
#> Boot 50

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

 res_val$intval

#>            Orig  Apparent      Test    Optimism Corrected
#> AUC   0.8631544 0.8722783 0.8493633  0.02291498 0.8402395
#> R2    0.4873082 0.5161438 0.4550536  0.06109022 0.4262179
#> Brier 0.1490050 0.1416810 0.1581931 -0.01651210 0.1655171
#> Slope 1.0000000 1.0000000 0.8730163  0.12698367 0.8730163

 res <- pool_intadj(res_psfmi, shrinkage_factor = 0.8730036)
 res$int_adj

#> [1] -5.088459

 res$coef_shrink_pooled

#>      Gender        Pain  Tampascale     Smoking    Function   Radiation 
#> -0.23322412  0.45829093  0.09188597  0.14637976 -0.05770793  0.23107101 
#>         Age 
#> -0.01377269