Function to pool NRI measures over Multiply Imputed datasets

pool_reclassification Function to pool categorical and continuous NRI and IDI over Multiply Imputed datasets

pool_reclassification(datasets, cutoff = cutoff)

Arguments

datasets
a list of data frames corresponding to the multiply imputed datasets, within each dataset in the first column the predicted probabilities of model 1, in the second column those of model 2 and in the third column the observed outcomes coded as '0'and '1'.

cutoff
cutoff value for the categorical NRI, must lie between 0 and 1.

Details

This function is called by the function pool_compare_model

Author

Martijn Heymans, 2020