Pooling and predictor selection function for Cox regression models in multiply imputed datasets

psfmi_coxr Pooling and backward selection for Cox regression models in multiply imputed datasets using different selection methods.

psfmi_coxr(
  data,
  nimp = 5,
  impvar = NULL,
  time,
  status,
  predictors = NULL,
  p.crit = 1,
  cat.predictors = NULL,
  spline.predictors = NULL,
  int.predictors = NULL,
  keep.predictors = NULL,
  knots = NULL,
  method = "RR",
  print.method = FALSE
)

Arguments

data
Data frame with stacked multiple imputed datasets. The original dataset that contains missing values must be excluded from the dataset. The imputed datasets must be distinguished by an imputation variable, specified under impvar, and starting by 1.

nimp
A numerical scalar. Number of imputed datasets. Default is 5.

impvar
A character vector. Name of the variable that distinguishes the imputed datasets.

time
Follow up time.

status
The status variable, normally 0=censoring, 1=event.

predictors
Character vector with the names of the predictor variables. At least one predictor variable has to be defined.

p.crit
A numerical scalar. P-value selection criterium.

cat.predictors
A single string or a vector of strings to define the categorical variables. Default is NULL categorical predictors.

spline.predictors
A single string or a vector of strings to define the (restricted cubic) spline variables. Default is NULL spline predictors.

int.predictors
A single string or a vector of strings with the names of the variables that form an interaction pair, separated by a “:” symbol.

keep.predictors
A single string or a vector of strings including the variables that are forced in the model during predictor selection. Categorical and interaction variables are allowed. See details.

knots
A numerical vector that defines the number of knots for each spline predictor separately.

method
A character vector to indicate the pooling method for p-values to pool the total model or used during predictor selection. This can be "D1", "D2", or "MPR". See details for more information.

print.method
logical vector. If TRUE full matrix with p-values of all variables according to chosen method (under method) is shown. If FALSE (default) p-value for categorical variables according to method are shown and for continuous and dichotomous predictors Rubin’s Rules are used.

Value

An object of class smodsmi (selected models in multiply imputed datasets) from which the following objects can be extracted: imputed datasets as data, selected pooled model as RR_model, pooled p-values according to pooling method as multiparm, predictors included at each selection step as predictors_in, predictors excluded at each step as predictors_out, and impvar, nimp, time, status, method, p.crit, predictors, cat.predictors, keep.predictors, int.predictors, spline.predictors, knots, print.method, call, model_type and predictors_final for names of predictors in final selection step, fit.formula is the regression formula of start model and predictors_initial for names of predictors in start model.

Details

The basic pooling procedure to derive pooled coefficients, standard errors, 95 confidence intervals and p-values is Rubin's Rules (RR). Specific procedures are available to derive pooled p-values for categorical (> 2 categories) and spline variables. print.method allows to choose between these pooling methods that are: “D1” is pooling of the total covariance matrix, ”D2” is pooling of Chi-square values, and “MPR” is pooling of median p-values (MPR rule). Spline regression coefficients are defined by using the rcs function for restricted cubic splines of the rms package of Frank Harrell. A minimum number of 3 knots as defined under knots is needed.

References

Eekhout I, van de Wiel MA, Heymans MW. Methods for significance testing of categorical covariates in logistic regression models after multiple imputation: power and applicability analysis. BMC Med Res Methodol. 2017;17(1):129.

Enders CK (2010). Applied missing data analysis. New York: The Guilford Press.

van de Wiel MA, Berkhof J, van Wieringen WN. Testing the prediction error difference between 2 predictors. Biostatistics. 2009;10:550-60.

Marshall A, Altman DG, Holder RL, Royston P. Combining estimates of interest in prognostic modelling studies after multiple imputation: current practice and guidelines. BMC Med Res Methodol. 2009;9:57.

Van Buuren S. (2018). Flexible Imputation of Missing Data. 2nd Edition. Chapman & Hall/CRC Interdisciplinary Statistics. Boca Raton.

http://missingdatasolutions.rbind.io/

Examples

  pool_coxr <- psfmi_coxr(data=lbpmicox, nimp=5, impvar="Impnr", time="Time",
  status="Status", predictors=c("Duration", "Radiation", "Onset"), p.crit=1,
  method="D1", cat.predictors=c("Expect_cat"))

#> 
#> Pooled model correctly estimated
#>             using a p-value of 1

  pool_coxr$RR_Model

#> $`Step 1`
#>                         est std.err signif   lower   upper     HR   L.HR   U.HR
#> Duration            -0.0074  0.0037 0.0467 -0.0147 -0.0001 0.9926 0.9854 0.9999
#> Radiation           -0.1675  0.1477 0.2569 -0.4570  0.1221 0.8458 0.6332 1.1299
#> Onset               -0.1065  0.1742 0.5411 -0.4480  0.2350 0.8990 0.6389 1.2649
#> factor(Expect_cat)2  0.1781  0.2244 0.4275 -0.2618  0.6180 1.1949 0.7697 1.8552
#> factor(Expect_cat)3  0.1718  0.1967 0.3823 -0.2137  0.5573 1.1874 0.8076 1.7460
#> 

  pool_coxr$multiparm

#> $`Step 1`
#>                    Chi_sq p-value D1 & RR
#> Duration            3.956          0.0467
#> Radiation           1.285          0.2569
#> Onset               0.373          0.5411
#> factor(Expect_cat)  0.420          0.6570
#> 

 if (FALSE) {
  pool_coxr <- psfmi_coxr(data=lbpmicox, nimp=5, impvar="Impnr", time="Time",
  status="Status", predictors=c("Previous",  "Radiation", "Onset",
  "Function", "Tampascale" ), p.crit=0.05, cat.predictors=c("Expect_cat"),
  int.predictors=c("Tampascale:Radiation",
  "Expect_cat:Tampascale"), keep.predictors = "Tampascale", method="D2")
  pool_coxr$RR_Model
  pool_coxr$multiparm
  pool_coxr$predictors_in
 }