Pooling and Predictor selection function for backward or forward selection of Linear regression models across multiply imputed data.

psfmi_lm Pooling and backward or forward selection of Linear regression models in multiply imputed data using selection methods RR, D1, D2 and MPR.

psfmi_lm(
  data,
  formula = NULL,
  nimp = 5,
  impvar = NULL,
  Outcome = NULL,
  predictors = NULL,
  cat.predictors = NULL,
  spline.predictors = NULL,
  int.predictors = NULL,
  keep.predictors = NULL,
  nknots = NULL,
  p.crit = 1,
  method = "RR",
  direction = NULL
)

Arguments

data

Data frame with stacked multiple imputed datasets. The original dataset that contains missing values must be excluded from the dataset. The imputed datasets must be distinguished by an imputation variable, specified under impvar, and starting by 1.

formula

A formula object to specify the model as normally used by glm. See under "Details" and "Examples" how these can be specified. If a formula object is used set predictors, cat.predictors, spline.predictors or int.predictors at the default value of NULL.

nimp

A numerical scalar. Number of imputed datasets. Default is 5.

impvar

A character vector. Name of the variable that distinguishes the imputed datasets.

Outcome

Character vector containing the name of the continuous outcome variable.

predictors

Character vector with the names of the predictor variables. At least one predictor variable has to be defined. Give predictors unique names and do not use predictor name combinations with numbers as, age2, gender10, etc.

cat.predictors

A single string or a vector of strings to define the categorical variables. Default is NULL categorical predictors.

spline.predictors

A single string or a vector of strings to define the (restricted cubic) spline variables. Default is NULL spline predictors. See details.

int.predictors

A single string or a vector of strings with the names of the variables that form an interaction pair, separated by a “:” symbol.

keep.predictors

A single string or a vector of strings including the variables that are forced in the model during predictor selection. All type of variables are allowed.

nknots

A numerical vector that defines the number of knots for each spline predictor separately.

p.crit

A numerical scalar. P-value selection criterium. A value of 1 provides the pooled model without selection.

method

A character vector to indicate the pooling method for p-values to pool the total model or used during predictor selection. This can be "RR", D1", "D2", "D3" or "MPR". See details for more information. Default is "RR".

direction

The direction of predictor selection, "BW" means backward selection and "FW" means forward selection.

Value

An object of class pmods (multiply imputed models) from which the following objects can be extracted:

data imputed datasets

RR_model pooled model at each selection step

RR_model_final final selected pooled model

multiparm pooled p-values at each step according to pooling method

multiparm_final pooled p-values at final step according to pooling method

multiparm_out (only when direction = "FW") pooled p-values of removed predictors

formula_step formula object at each step

formula_final formula object at final step

formula_initial formula object at final step

predictors_in predictors included at each selection step

predictors_out predictors excluded at each step

impvar name of variable used to distinguish imputed datasets

nimp number of imputed datasets

Outcome name of the outcome variable

method selection method

p.crit p-value selection criterium

call function call

model_type type of regression model used

direction direction of predictor selection

predictors_final names of predictors in final selection step

predictors_initial names of predictors in start model

keep.predictors names of predictors that were forced in the model

Details

The basic pooling procedure to derive pooled coefficients, standard errors, 95 confidence intervals and p-values is Rubin's Rules (RR). However, RR is only possible when the model included continuous or dichotomous variables. Specific procedures are available when the model also included categorical (> 2 categories) or restricted cubic spline variables. These pooling methods are: “D1” is pooling of the total covariance matrix, ”D2” is pooling of Chi-square values and “MPR” is pooling of median p-values (MPR rule). Spline regression coefficients are defined by using the rcs function for restricted cubic splines of the rms package. A minimum number of 3 knots as defined under knots is required.

A typical formula object has the form Outcome ~ terms. Categorical variables has to be defined as Outcome ~ factor(variable), restricted cubic spline variables as Outcome ~ rcs(variable, 3). Interaction terms can be defined as Outcome ~ variable1*variable2 or Outcome ~ variable1 + variable2 + variable1:variable2. All variables in the terms part have to be separated by a "+". If a formula object is used set predictors, cat.predictors, spline.predictors or int.predictors at the default value of NULL.

References

Enders CK (2010). Applied missing data analysis. New York: The Guilford Press.

van de Wiel MA, Berkhof J, van Wieringen WN. Testing the prediction error difference between 2 predictors. Biostatistics. 2009;10:550-60.

Marshall A, Altman DG, Holder RL, Royston P. Combining estimates of interest in prognostic modelling studies after multiple imputation: current practice and guidelines. BMC Med Res Methodol. 2009;9:57.

Van Buuren S. (2018). Flexible Imputation of Missing Data. 2nd Edition. Chapman & Hall/CRC Interdisciplinary Statistics. Boca Raton.

EW. Steyerberg (2019). Clinical Prediction MOdels. A Practical Approach to Development, Validation, and Updating (2nd edition). Springer Nature Switzerland AG.

http://missingdatasolutions.rbind.io/

Author

Martijn Heymans, 2021

Examples

  pool_lm <- psfmi_lm(data=lbpmilr, formula = Pain ~ factor(Satisfaction) + 
  rcs(Tampascale,3) + Radiation + 
  Radiation*factor(Satisfaction) + Age + Duration + BMI,
  p.crit = 0.05, direction="FW", nimp=5, impvar="Impnr", 
  keep.predictors = c("Radiation*factor(Satisfaction)", "Age"), method="D1")
#> Entered at Step 1 is - rcs(Tampascale,3)
#> 
#> Selection correctly terminated, 
#> No new variables entered the model
  
  pool_lm$RR_model_final
#> $`Final model`
#>                              term     estimate  std.error   statistic        df
#> 1                     (Intercept) -5.962723799 1.49224638 -3.99580382  58.68977
#> 2    rcs(Tampascale, 3)Tampascale  0.273576706 0.04090154  6.68866561  80.20178
#> 3   rcs(Tampascale, 3)Tampascale'  0.068536039 0.05519202  1.24177437  54.33244
#> 4                       Radiation -0.307493209 0.45522228 -0.67547925  20.79741
#> 5                             Age  0.003684182 0.01118725  0.32931959  25.74372
#> 6           factor(Satisfaction)2 -0.011193910 0.27056950 -0.04137166 127.01320
#> 7           factor(Satisfaction)3  0.590228529 0.53933762  1.09435817  13.75653
#> 8 Radiation:factor(Satisfaction)2  0.364188395 0.48755421  0.74697005  38.86693
#> 9 Radiation:factor(Satisfaction)3  0.060103139 0.76466224  0.07860090  11.02087
#>        p.value
#> 1 1.823215e-04
#> 2 2.769264e-09
#> 3 2.196537e-01
#> 4 5.068130e-01
#> 5 7.445785e-01
#> 6 9.670646e-01
#> 7 2.925899e-01
#> 8 4.595768e-01
#> 9 9.387591e-01
#>