Function to evaluate bootstrap predictor and model stability in multiply imputed datasets.

psfmi_stab Stability analysis of predictors and prediction models selected with the psfmi_lr, psfmi_coxr or psfmi_mm functions of the psfmi package.

psfmi_stab(pobj, boot_method = NULL, nboot = 20)

Arguments

pobj
An object of class smodsmi (selected models in multiply imputed datasets), produced by a previous call to psfmi_lr, psfmi_coxr or psfmi_mm.

boot_method
A single string to define the bootstrap method. Use "single" after a call to psfmi_lr and psfmi_coxr and "cluster" after a call to psfmi_mm.

nboot
A numerical scalar. Number of bootstrap samples to evaluate the stability. Default is 20.

Value

A psfmi_stab object from which the following objects can be extracted: bootstrap inclusion (selection) frequency of each predictor bif, total number each predictor is included in the bootstrap samples as bif_total, percentage a predictor is selected in each bootstrap sample as bif_perc and number of times a prediction model is selected in the bootstrap samples as model_stab.

Details

The function evaluates predictor selection frequency in stratified or cluster bootstrap samples. The stratification factor is the variable that separates the imputed datasets. It uses as input an object of class smodsmi as a result of a previous call to the psfmi_lr, psfmi_coxr or psfmi_mm functions. In combination with the psfmi_mm function a cluster bootstrap method is used where bootstrapping is used on the level of the clusters only.

References

Heymans MW, van Buuren S. et al. Variable selection under multiple imputation using the bootstrap in a prognostic study. BMC Med Res Methodol. 2007;13:7-33.

Eekhout I, van de Wiel MA, Heymans MW. Methods for significance testing of categorical covariates in logistic regression models after multiple imputation: power and applicability analysis. BMC Med Res Methodol. 2017;17(1):129.

Sauerbrei W, Schumacher M. A bootstrap resampling procedure for model building: application to the Cox regression model. Stat Med. 1992;11:2093–109.

Royston P, Sauerbrei W (2008) Multivariable model-building – a pragmatic approach to regression analysis based on fractional polynomials for modelling continuous variables. (2008). Chapter 8, Model Stability. Wiley, Chichester

Heinze G, Wallisch C, Dunkler D. Variable selection - A review and recommendations for the practicing statistician. Biom J. 2018;60(3):431-449.

http://missingdatasolutions.rbind.io/

Examples

 pool_lr <- psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
                   predictors=c("Gender", "Smoking",  "JobControl", "JobDemands",
                   "Age", "Radiation", "SocialSupport", "Function"),
                   cat.predictors = c("Carrying"), p.crit =0.157, method="D1")

#> Variable excluded at Step 1 is - JobDemands

#> Variable excluded at Step 2 is - Smoking

#> Variable excluded at Step 3 is - Gender

#> Variable excluded at Step 4 is - JobControl

#> Variable excluded at Step 5 is - SocialSupport

#> Variable excluded at Step 6 is - Function

 pool_lr$RR_Model

#> $`Step 1`
#>                       est std.err signif   lower  upper     OR   L.OR    U.OR
#> (Intercept)       -1.6939  2.9211 0.5620 -7.4220 4.0341 0.1838 0.0006 56.4915
#> Gender            -0.2009  0.4780 0.6742 -1.1381 0.7362 0.8180 0.3204  2.0879
#> Smoking            0.1234  0.3802 0.7456 -0.6218 0.8685 1.1313 0.5370  2.3834
#> JobControl         0.0115  0.0228 0.6152 -0.0335 0.0565 1.0115 0.9670  1.0581
#> JobDemands        -0.0086  0.0421 0.8384 -0.0912 0.0740 0.9914 0.9128  1.0768
#> Age               -0.0286  0.0215 0.1847 -0.0708 0.0137 0.9718 0.9316  1.0138
#> Radiation          1.0647  0.3946 0.0071  0.2899 1.8394 2.8998 1.3363  6.2926
#> SocialSupport      0.0629  0.0647 0.3315 -0.0642 0.1901 1.0649 0.9378  1.2093
#> Function          -0.0646  0.0509 0.2052 -0.1645 0.0354 0.9375 0.8483  1.0361
#> factor(Carrying)2  1.1026  0.5179 0.0334  0.0870 2.1181 3.0119 1.0909  8.3152
#> factor(Carrying)3  1.8879  0.5641 0.0009  0.7800 2.9958 6.6054 2.1814 20.0015
#> 
#> $`Step 2`
#>                       est std.err signif   lower  upper     OR   L.OR    U.OR
#> (Intercept)       -2.0409  2.5225 0.4194 -7.0157 2.9338 0.1299 0.0009 18.7988
#> Gender            -0.1829  0.4694 0.6968 -1.1030 0.7371 0.8328 0.3319  2.0899
#> Smoking            0.1214  0.3802 0.7496 -0.6239 0.8667 1.1291 0.5359  2.3790
#> JobControl         0.0117  0.0228 0.6098 -0.0335 0.0569 1.0118 0.9671  1.0585
#> Age               -0.0280  0.0215 0.1939 -0.0702 0.0143 0.9724 0.9322  1.0144
#> Radiation          1.0539  0.3911 0.0072  0.2860 1.8217 2.8687 1.3311  6.1824
#> SocialSupport      0.0629  0.0646 0.3305 -0.0640 0.1899 1.0650 0.9380  1.2091
#> Function          -0.0643  0.0510 0.2079 -0.1645 0.0359 0.9377 0.8483  1.0365
#> factor(Carrying)2  1.1122  0.5189 0.0323  0.0943 2.1300 3.0409 1.0989  8.4150
#> factor(Carrying)3  1.8946  0.5659 0.0009  0.7826 3.0066 6.6499 2.1871 20.2189
#> 
#> $`Step 3`
#>                       est std.err signif   lower  upper     OR   L.OR    U.OR
#> (Intercept)       -2.0332  2.5252 0.4217 -7.0141 2.9477 0.1309 0.0009 19.0629
#> Gender            -0.1707  0.4669 0.7146 -1.0860 0.7445 0.8430 0.3376  2.1054
#> JobControl         0.0119  0.0228 0.6019 -0.0332 0.0571 1.0120 0.9673  1.0587
#> Age               -0.0276  0.0215 0.1998 -0.0700 0.0147 0.9727 0.9324  1.0148
#> Radiation          1.0486  0.3893 0.0072  0.2844 1.8128 2.8536 1.3290  6.1273
#> SocialSupport      0.0627  0.0646 0.3323 -0.0642 0.1896 1.0647 0.9378  1.2087
#> Function          -0.0637  0.0507 0.2097 -0.1632 0.0359 0.9383 0.8494  1.0366
#> factor(Carrying)2  1.1353  0.5127 0.0269  0.1298 2.1409 3.1122 1.1386  8.5070
#> factor(Carrying)3  1.8966  0.5649 0.0009  0.7865 3.0066 6.6631 2.1958 20.2191
#> 
#> $`Step 4`
#>                       est std.err signif   lower  upper     OR   L.OR    U.OR
#> (Intercept)       -2.2583  2.4678 0.3616 -7.1341 2.6174 0.1045 0.0008 13.7000
#> JobControl         0.0136  0.0225 0.5480 -0.0311 0.0582 1.0137 0.9694  1.0599
#> Age               -0.0287  0.0212 0.1764 -0.0703 0.0130 0.9717 0.9321  1.0130
#> Radiation          1.0504  0.3883 0.0070  0.2883 1.8124 2.8587 1.3342  6.1252
#> SocialSupport      0.0651  0.0642 0.3115 -0.0611 0.1912 1.0672 0.9408  1.2107
#> Function          -0.0662  0.0500 0.1856 -0.1644 0.0319 0.9359 0.8484  1.0324
#> factor(Carrying)2  1.1718  0.5035 0.0201  0.1840 2.1596 3.2278 1.2020  8.6673
#> factor(Carrying)3  1.8997  0.5645 0.0008  0.7904 3.0090 6.6838 2.2043 20.2662
#> 
#> $`Step 5`
#>                       est std.err signif   lower  upper     OR   L.OR    U.OR
#> (Intercept)       -1.3222  1.8561 0.4767 -4.9722 2.3278 0.2665 0.0069 10.2554
#> Age               -0.0291  0.0212 0.1705 -0.0707 0.0126 0.9713 0.9317  1.0126
#> Radiation          1.0216  0.3796 0.0072  0.2769 1.7662 2.7776 1.3191  5.8487
#> SocialSupport      0.0571  0.0625 0.3616 -0.0657 0.1799 1.0588 0.9364  1.1971
#> Function          -0.0614  0.0500 0.2206 -0.1597 0.0370 0.9405 0.8524  1.0377
#> factor(Carrying)2  1.1826  0.5027 0.0188  0.1964 2.1689 3.2630 1.2170  8.7486
#> factor(Carrying)3  1.9089  0.5662 0.0008  0.7955 3.0223 6.7457 2.2156 20.5382
#> 
#> $`Step 6`
#>                       est std.err signif   lower  upper     OR   L.OR    U.OR
#> (Intercept)        0.0100  1.1407 0.9930 -2.2399 2.2599 1.0101 0.1065  9.5825
#> Age               -0.0297  0.0213 0.1634 -0.0716 0.0122 0.9707 0.9309  1.0122
#> Radiation          1.0145  0.3792 0.0076  0.2706 1.7584 2.7579 1.3107  5.8032
#> Function          -0.0580  0.0500 0.2467 -0.1564 0.0404 0.9436 0.8552  1.0412
#> factor(Carrying)2  1.1744  0.5026 0.0196  0.1884 2.1604 3.2362 1.2073  8.6749
#> factor(Carrying)3  1.8784  0.5656 0.0010  0.7660 2.9908 6.5429 2.1511 19.9008
#> 
#> $`Step 7`
#>                       est std.err signif   lower  upper     OR   L.OR    U.OR
#> (Intercept)       -0.7085  0.9443 0.4538 -2.5690 1.1520 0.4924 0.0766  3.1645
#> Age               -0.0309  0.0210 0.1428 -0.0723 0.0105 0.9696 0.9303  1.0106
#> Radiation          1.0319  0.3754 0.0060  0.2957 1.7682 2.8065 1.3440  5.8605
#> factor(Carrying)2  1.2860  0.4886 0.0085  0.3279 2.2441 3.6182 1.3880  9.4319
#> factor(Carrying)3  2.1525  0.5058 0.0000  1.1601 3.1449 8.6066 3.1904 23.2176
#> 

 pool_lr$multiparm

#> $`Step 1`
#>                  Chi_sq p-value D1 & RR
#> Gender            0.177          0.6742
#> Smoking           0.105          0.7456
#> JobControl        0.254          0.6152
#> JobDemands        0.042          0.8384
#> Age               1.765          0.1847
#> Radiation         7.280          0.0071
#> SocialSupport     0.944          0.3315
#> Function          1.609          0.2052
#> factor(Carrying)  5.724          0.0035
#> 
#> $`Step 2`
#>                  Chi_sq p-value D1 & RR
#> Gender            0.152          0.6968
#> Smoking           0.102          0.7496
#> JobControl        0.262          0.6098
#> Age               1.695          0.1939
#> Radiation         7.262          0.0072
#> SocialSupport     0.949          0.3305
#> Function          1.590          0.2079
#> factor(Carrying)  5.720          0.0035
#> 
#> $`Step 3`
#>                  Chi_sq p-value D1 & RR
#> Gender            0.134          0.7146
#> JobControl        0.273          0.6019
#> Age               1.651          0.1998
#> Radiation         7.255          0.0072
#> SocialSupport     0.942          0.3323
#> Function          1.577          0.2097
#> factor(Carrying)  5.765          0.0034
#> 
#> $`Step 4`
#>                  Chi_sq p-value D1 & RR
#> JobControl        0.363          0.5480
#> Age               1.835          0.1764
#> Radiation         7.319          0.0070
#> SocialSupport     1.026          0.3115
#> Function          1.756          0.1856
#> factor(Carrying)  5.790          0.0033
#> 
#> $`Step 5`
#>                  Chi_sq p-value D1 & RR
#> Age               1.887          0.1705
#> Radiation         7.241          0.0072
#> SocialSupport     0.833          0.3616
#> Function          1.506          0.2206
#> factor(Carrying)  5.840          0.0032
#> 
#> $`Step 6`
#>                  Chi_sq p-value D1 & RR
#> Age               1.955          0.1634
#> Radiation         7.157          0.0076
#> Function          1.347          0.2467
#> factor(Carrying)  5.684          0.0037
#> 
#> $`Step 7`
#>                  Chi_sq p-value D1 & RR
#> Age               2.161          0.1428
#> Radiation         7.558          0.0060
#> factor(Carrying)  9.166          0.0001
#> 

if (FALSE) {
 stab_res <- psfmi_stab(pool_lr, boot_method = "single", nboot=50)
 stab_res$bif
 stab_res$bif_perc
 stab_res$model_stab
}