Provides pooled adjusted intercept after shrinkage of pooled coefficients in multiply imputed datasets

pool_intadj Provides pooled adjusted intercept after shrinkage of the pooled coefficients in multiply imputed datasets for models selected with the psfmi_lr function and internally validated with the psfmi_perform function.

pool_intadj(pobj, shrinkage_factor)

Arguments

pobj
An object of class smodsmi (selected models in multiply imputed datasets), produced by a previous call to psfmi_lr.

shrinkage_factor
A numerical scalar. Shrinkage factor value as a result of internal validation with the psfmi_perform function.

Value

A pool_intadj object from which the following objects can be extracted: int_adj, the adjusted intercept value, coef_shrink_pooled, the pooled regression coefficients after shrinkage, coef_orig_pooled, the (original) pooled regression coefficients before shrinkage and nimp, the number of imputed datasets.

Details

The function provides the pooled adjusted intercept after shrinkage of pooled regression coefficients in multiply imputed datasets. The function is only available for logistic regression models without random effects.

References

F. Harrell. Regression Modeling Strategies. With Applications to Linear Models, Logistic and Ordinal Regression, and Survival Analysis (2nd edition). Springer, New York, NY, 2015.

EW. Steyerberg (2019). Clinical Prediction MOdels. A Practical Approach to Development, Validation, and Updating (2nd edition). Springer Nature Switzerland AG.

http://missingdatasolutions.rbind.io/

Examples

 res_psfmi <- psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
           predictors=c("Gender", "Pain","Tampascale","Smoking","Function",
           "Radiation", "Age"), p.crit = 1, method="D1")

#> Registered S3 methods overwritten by 'car':
#>   method                          from
#>   influence.merMod                lme4
#>   cooks.distance.influence.merMod lme4
#>   dfbeta.influence.merMod         lme4
#>   dfbetas.influence.merMod        lme4

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

 res_psfmi$RR_Model

#> $`Step 1`
#>                 est std.err signif    lower  upper     OR   L.OR   U.OR
#> (Intercept) -5.7989  3.0757 0.0630 -11.9196 0.3218 0.0030 0.0000 1.3797
#> Gender      -0.2672  0.5041 0.5962  -1.2554 0.7211 0.7656 0.2850 2.0567
#> Pain         0.5250  0.2396 0.0325   0.0453 1.0046 1.6904 1.0464 2.7308
#> Tampascale   0.1053  0.0905 0.2485  -0.0751 0.2856 1.1110 0.9277 1.3306
#> Smoking      0.1677  0.4236 0.6923  -0.6629 0.9983 1.1826 0.5153 2.7136
#> Function    -0.0661  0.0523 0.2062  -0.1686 0.0364 0.9360 0.8448 1.0371
#> Radiation    0.2647  0.4685 0.5725  -0.6572 1.1866 1.3030 0.5183 3.2758
#> Age         -0.0158  0.0235 0.5015  -0.0618 0.0303 0.9843 0.9400 1.0307
#> 

if (FALSE) {
 set.seed(100)
 res_val <- psfmi_perform(res_psfmi, method = "MI_boot", nboot=10,
   int_val = TRUE, p.crit=1, cal.plot=FALSE, plot.indiv=FALSE)
 res_val$intval

 res <- pool_intadj(res_psfmi, shrinkage_factor = 0.9774058)
 res$int_adj
 res$coef_shrink_pooled
 }