Pooling and predictor selection function for Cox regression models in multiply imputed datasets

psfmi_coxr Pooling and backward selection for Cox regression models in multiply imputed datasets using different selection methods.

psfmi_coxr(data, nimp = 5, impvar = NULL, time, status,
  predictors = NULL, p.crit = 1, cat.predictors = NULL,
  spline.predictors = NULL, int.predictors = NULL,
  keep.predictors = NULL, knots = NULL, method = NULL,
  print.method = FALSE)

Arguments

data
Data frame or data matrix with stacked multiple imputed datasets. The original dataset that contains missing values must be excluded from the dataset. The imputed datasets must be distinguished by an imputation variable, specified under impvar, and starting by 1.

nimp
A numerical scalar. Number of imputed datasets. Default is 5.

impvar
A character vector. Name of the variable that distinguishes the imputed datasets.

time
Follow up time.

status
The status variable, normally 0=censoring, 1=event.

predictors
Character vector with the names of the predictor variables. At least one predictor variable has to be defined.

p.crit
A numerical scalar. P-value selection criterium.

cat.predictors
A single string or a vector of strings to define the categorical variables. Default is NULL categorical predictors.

spline.predictors
A single string or a vector of strings to define the (restricted cubic) spline variables. Default is NULL spline predictors.

int.predictors
A single string or a vector of strings with the names of the variables that form an interaction pair, separated by a “:” symbol.

keep.predictors
A single string or a vector of strings including the variables that are forced in the model during predictor selection. Categorical and interaction variables are allowed. See details.

knots
A numerical vector that defines the number of knots for each spline predictor separately.

method
A character vector to indicate the pooling method for p-values to pool the total model or used during predictor selection. This can be "D1", "D2", or "MPR". See details for more information.

print.method
logical vector. If TRUE full matrix with p-values of all variables according to chosen method (under method) is shown. If FALSE (default) p-value for categorical variables according to method are shown and for continuous and dichotomous predictors Rubin’s Rules are used.

Value

A psfmi_coxr object from which the following objects can be extracted: pooled model as RR_model, pooled p-values according to pooling method as multiparm_p, predictors excluded at each step as coef.excl_step, and impvar, nimp, method, p.crit, predictors, cat.predictors, keep.predictors, int.predictors, spline.predictors, knots, print.method.

Details

The basic pooling procedure to derive pooled coefficients, standard errors, 95 confidence intervals and p-values is Rubin's Rules (RR). Specific procedures are available to derive pooled p-values for categorical (> 2 categories) and spline variables. print.method allows to choose between these pooling methods that are: “D1” is pooling of the total covariance matrix, ”D2” is pooling of Chi-square values, and “MPR” is pooling of median p-values (MPR rule). Spline regression coefficients are defined by using the rcs function for restricted cubic splines of the rms package of Frank Harrell. A minimum number of 3 knots as defined under knots is needed.

References

Eekhout I, van de Wiel MA, Heymans MW. Methods for significance testing of categorical covariates in logistic regression models after multiple imputation: power and applicability analysis. BMC Med Res Methodol. 2017;17(1):129.

Enders CK (2010). Applied missing data analysis. New York: The Guilford Press.

van de Wiel MA, Berkhof J, van Wieringen WN. Testing the prediction error difference between 2 predictors. Biostatistics. 2009;10:550-60.

Marshall A, Altman DG, Holder RL, Royston P. Combining estimates of interest in prognostic modelling studies after multiple imputation: current practice and guidelines. BMC Med Res Methodol. 2009;9:57.

Van Buuren S. (2018). Flexible Imputation of Missing Data. 2nd Edition. Chapman & Hall/CRC Interdisciplinary Statistics. Boca Raton.

http://missingdatasolutions.rbind.io/

Examples

  pool_coxr <- psfmi_coxr(data=lbpmicox, nimp=5, impvar="Impnr", time="Time",
  status="Status", predictors=c("Duration", "Radiation", "Onset"), p.crit=1,
  method="D1", cat.predictors=c("Expect_cat"))

#> 
#> Pooled model correctly estimated
#>             using a p-value of 1

  pool_coxr$RR_Model

#> $`Step 1`
#>                         est std.err signif   lower   upper     HR   L.HR   U.HR
#> Duration            -0.0074  0.0037 0.0467 -0.0147 -0.0001 0.9926 0.9854 0.9999
#> Radiation           -0.1675  0.1477 0.2569 -0.4570  0.1221 0.8458 0.6332 1.1299
#> Onset               -0.1065  0.1742 0.5411 -0.4480  0.2350 0.8990 0.6389 1.2649
#> factor(Expect_cat)2  0.1781  0.2244 0.4275 -0.2618  0.6180 1.1949 0.7697 1.8552
#> factor(Expect_cat)3  0.1718  0.1967 0.3823 -0.2137  0.5573 1.1874 0.8076 1.7460
#> 

  pool_coxr$multiparm_p

#> $`Step 1`
#>                    Chi_sq D1 & RR p-values
#> Duration            3.956           0.0467
#> Radiation           1.285           0.2569
#> Onset               0.373           0.5411
#> factor(Expect_cat)  0.420           0.6570
#> 

  pool_coxr <- psfmi_coxr(data=lbpmicox, nimp=5, impvar="Impnr", time="Time",
  status="Status", predictors=c("Previous",  "Radiation", "Onset",
  "Function", "Tampascale" ), p.crit=0.05, cat.predictors=c("Expect_cat"),
  int.predictors=c("Tampascale:Radiation",
  "Expect_cat:Tampascale"), keep.predictors = "Tampascale", method="D2")

#> Variable excluded at Step 1 is - Previous

#> Variable excluded at Step 2 is - Onset

#> 
#> Pooled model correctly estimated
#>           using a p-value of 0.05 and predictors to keep Tampascale

  pool_coxr$RR_Model

#> $`Step 1`
#>                                    est std.err signif   lower   upper     HR
#> Previous                       -0.1386  0.1951 0.4773 -0.5210  0.2438 0.8706
#> Radiation                      -1.6289  1.0685 0.1276 -3.7245  0.4668 0.1961
#> Onset                          -0.1440  0.1769 0.4159 -0.4908  0.2029 0.8659
#> Function                        0.0436  0.0176 0.0133  0.0091  0.0781 1.0446
#> Tampascale                     -0.0825  0.0371 0.0262 -0.1552 -0.0098 0.9208
#> factor(Expect_cat)2            -2.3433  1.5850 0.1394 -5.4514  0.7648 0.0960
#> factor(Expect_cat)3            -1.6527  1.4176 0.2437 -4.4312  1.1258 0.1915
#> Radiation:Tampascale            0.0378  0.0262 0.1504 -0.0137  0.0892 1.0385
#> Tampascale:factor(Expect_cat)2  0.0632  0.0396 0.1111 -0.0145  0.1409 1.0652
#> Tampascale:factor(Expect_cat)3  0.0467  0.0364 0.1995 -0.0246  0.1180 1.0478
#>                                  L.HR   U.HR
#> Previous                       0.5939 1.2761
#> Radiation                      0.0241 1.5949
#> Onset                          0.6121 1.2249
#> Function                       1.0091 1.0812
#> Tampascale                     0.8562 0.9902
#> factor(Expect_cat)2            0.0043 2.1486
#> factor(Expect_cat)3            0.0119 3.0827
#> Radiation:Tampascale           0.9864 1.0933
#> Tampascale:factor(Expect_cat)2 0.9856 1.1513
#> Tampascale:factor(Expect_cat)3 0.9757 1.1252
#> 
#> $`Step 2`
#>                                    est std.err signif   lower   upper     HR
#> Radiation                      -1.6501  1.0730 0.1243 -3.7545  0.4544 0.1920
#> Onset                          -0.1345  0.1764 0.4458 -0.4804  0.2113 0.8742
#> Function                        0.0435  0.0175 0.0132  0.0091  0.0779 1.0445
#> Tampascale                     -0.0826  0.0370 0.0257 -0.1551 -0.0100 0.9207
#> factor(Expect_cat)2            -2.3808  1.5765 0.1311 -5.4717  0.7102 0.0925
#> factor(Expect_cat)3            -1.6784  1.4177 0.2364 -4.4570  1.1002 0.1867
#> Radiation:Tampascale            0.0385  0.0263 0.1443 -0.0132  0.0902 1.0393
#> Tampascale:factor(Expect_cat)2  0.0640  0.0394 0.1043 -0.0132  0.1413 1.0661
#> Tampascale:factor(Expect_cat)3  0.0477  0.0364 0.1897 -0.0236  0.1189 1.0489
#>                                  L.HR   U.HR
#> Radiation                      0.0234 1.5752
#> Onset                          0.6185 1.2353
#> Function                       1.0091 1.0810
#> Tampascale                     0.8563 0.9900
#> factor(Expect_cat)2            0.0042 2.0344
#> factor(Expect_cat)3            0.0116 3.0048
#> Radiation:Tampascale           0.9869 1.0944
#> Tampascale:factor(Expect_cat)2 0.9869 1.1518
#> Tampascale:factor(Expect_cat)3 0.9767 1.1263
#> 
#> $`Step 3`
#>                                    est std.err signif   lower   upper     HR
#> Radiation                      -1.5369  1.0575 0.1463 -3.6109  0.5372 0.2150
#> Function                        0.0439  0.0176 0.0124  0.0095  0.0783 1.0449
#> Tampascale                     -0.0829  0.0369 0.0247 -0.1552 -0.0105 0.9204
#> factor(Expect_cat)2            -2.4364  1.5707 0.1210 -5.5161  0.6433 0.0875
#> factor(Expect_cat)3            -1.7333  1.4142 0.2203 -4.5050  1.0384 0.1767
#> Radiation:Tampascale            0.0357  0.0260 0.1690 -0.0152  0.0867 1.0363
#> Tampascale:factor(Expect_cat)2  0.0657  0.0393 0.0946 -0.0113  0.1427 1.0679
#> Tampascale:factor(Expect_cat)3  0.0492  0.0363 0.1751 -0.0219  0.1202 1.0504
#>                                  L.HR   U.HR
#> Radiation                      0.0270 1.7112
#> Function                       1.0095 1.0814
#> Tampascale                     0.8562 0.9896
#> factor(Expect_cat)2            0.0040 1.9027
#> factor(Expect_cat)3            0.0111 2.8247
#> Radiation:Tampascale           0.9849 1.0906
#> Tampascale:factor(Expect_cat)2 0.9888 1.1534
#> Tampascale:factor(Expect_cat)3 0.9783 1.1277
#> 

  pool_coxr$multiparm_p

#> $`Step 1`
#>                                   D2 D2 & RR p-values
#> Previous                      0.4910           0.4773
#> Radiation                     0.4915           0.1276
#> Onset                         0.6449           0.4159
#> Function                      5.7739           0.0133
#> Tampascale                    1.4903           0.0262
#> factor(Expect_cat)            0.3761           0.6866
#> Radiation:Tampascale          2.0270           0.1504
#> Tampascale:factor(Expect_cat) 1.2615           0.2834
#> 
#> $`Step 2`
#>                                   D2 D2 & RR p-values
#> Radiation                     0.4182           0.1243
#> Onset                         0.5670           0.4458
#> Function                      5.7778           0.0132
#> Tampascale                    1.3377           0.0257
#> factor(Expect_cat)            0.4344           0.6476
#> Radiation:Tampascale          2.0887           0.1443
#> Tampascale:factor(Expect_cat) 1.3182           0.2677
#> 
#> $`Step 3`
#>                                   D2 D2 & RR p-values
#> Radiation                     0.4040           0.1463
#> Function                      5.8834           0.0124
#> Tampascale                    1.4184           0.0247
#> factor(Expect_cat)            0.4620           0.6300
#> Radiation:Tampascale          1.8456           0.1690
#> Tampascale:factor(Expect_cat) 1.3962           0.2476
#>