Pooling and Predictor selection function for Logistic regression models in multiply imputed datasets

psfmi_lr Pooling and backward selection for Logistic regression prediction models in multiply imputed datasets using different selection methods.

psfmi_lr(data, nimp = 5, impvar = NULL, Outcome, predictors = NULL,
  p.crit = 1, cat.predictors = NULL, spline.predictors = NULL,
  int.predictors = NULL, keep.predictors = NULL, knots = NULL,
  method = NULL, print.method = FALSE)

Arguments

data
Data frame or data matrix with stacked multiple imputed datasets. The original dataset that contains missing values must be excluded from the dataset. The imputed datasets must be distinguished by an imputation variable, specified under impvar, and starting by 1.

nimp
A numerical scalar. Number of imputed datasets. Default is 5.

impvar
A character vector. Name of the variable that distinguishes the imputed datasets.

Outcome
Character vector containing the name of the outcome variable.

predictors
Character vector with the names of the predictor variables. At least one predictor variable has to be defined.

p.crit
A numerical scalar. P-value selection criterium.

cat.predictors
A single string or a vector of strings to define the categorical variables. Default is NULL categorical predictors.

spline.predictors
A single string or a vector of strings to define the (restricted cubic) spline variables. Default is NULL spline predictors. See details.

int.predictors
A single string or a vector of strings with the names of the variables that form an interaction pair, separated by a “:” symbol.

keep.predictors
A single string or a vector of strings including the variables that are forced in the model during predictor selection. Categorical and interaction variables are allowed.

knots
A numerical vector that defines the number of knots for each spline predictor separately.

method
A character vector to indicate the pooling method for p-values to pool the total model or used during predictor selection. This can be "D1", "D2", "D3" or "MPR". See details for more information.

print.method
logical vector. If TRUE full matrix with p-values of all variables according to chosen method (under method) is shown. If FALSE (default) p-value for categorical variables according to method are shown and for continuous and dichotomous predictors Rubin’s Rules are used.

Value

A psfmi_lr object from which the following objects can be extracted: pooled model as RR_model, pooled p-values according to pooling method as multiparm_p, predictors excluded at each step as coef.excl_step, and impvar, nimp, Outcome, method, p.crit, predictors, cat.predictors, keep.predictors, int.predictors, spline.predictors, knots, print.method.

Details

The basic pooling procedure to derive pooled coefficients, standard errors, 95 confidence intervals and p-values is Rubin's Rules (RR). Specific procedures are available to derive pooled p-values for categorical (> 2 categories) and spline variables. print.method allows to choose between these pooling methods that are: “D1” is pooling of the total covariance matrix, ”D2” is pooling of Chi-square values, “D3” is pooling Likelihood ratio statistics (method of Meng and Rubin) and “MPR” is pooling of median p-values (MPR rule). Spline regression coefficients are defined by using the rcs function for restricted cubic splines of the rms package of Frank Harrell. A minimum number of 3 knots as defined under knots is needed.

References

Eekhout I, van de Wiel MA, Heymans MW. Methods for significance testing of categorical covariates in logistic regression models after multiple imputation: power and applicability analysis. BMC Med Res Methodol. 2017;17(1):129.

Enders CK (2010). Applied missing data analysis. New York: The Guilford Press.

Meng X-L, Rubin DB. Performing likelihood ratio tests with multiply-imputed data sets. Biometrika.1992;79:103-11.

van de Wiel MA, Berkhof J, van Wieringen WN. Testing the prediction error difference between 2 predictors. Biostatistics. 2009;10:550-60.

Marshall A, Altman DG, Holder RL, Royston P. Combining estimates of interest in prognostic modelling studies after multiple imputation: current practice and guidelines. BMC Med Res Methodol. 2009;9:57.

Van Buuren S. (2018). Flexible Imputation of Missing Data. 2nd Edition. Chapman & Hall/CRC Interdisciplinary Statistics. Boca Raton.

http://missingdatasolutions.rbind.io/

Examples

  pool_lr <- psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
  predictors=c("Gender", "Smoking", "Function", "JobControl",
  "JobDemands", "SocialSupport"), method="D1")

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

  pool_lr$RR_Model

#> $`Step 1`
#>                   est std.err signif   lower   upper     OR   L.OR     U.OR
#> (Intercept)    1.1637  2.2828 0.6102 -3.3104  5.6379 3.2018 0.0365 280.8649
#> Gender        -0.4220  0.4073 0.3001 -1.2202  0.3762 0.6557 0.2952   1.4568
#> Smoking        0.1164  0.3407 0.7327 -0.5515  0.7842 1.1234 0.5761   2.1907
#> Function      -0.1457  0.0421 0.0005 -0.2282 -0.0632 0.8644 0.7959   0.9387
#> JobControl     0.0021  0.0191 0.9145 -0.0355  0.0396 1.0021 0.9652   1.0404
#> JobDemands    -0.0087  0.0357 0.8065 -0.0787  0.0612 0.9913 0.9243   1.0632
#> SocialSupport  0.0249  0.0540 0.6449 -0.0810  0.1308 1.0252 0.9222   1.1397
#> 

  pool_lr$multiparm_p

#> $`Step 1`
#>               Chi_sq D1 & RR p-values
#> Gender         1.074           0.3001
#> Smoking        0.117           0.7327
#> Function      11.984           0.0005
#> JobControl     0.012           0.9145
#> JobDemands     0.060           0.8065
#> SocialSupport  0.212           0.6449
#> 

  pool_lr <- psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
  predictors=c("Gender", "Smoking", "Function", "JobControl",
  "JobDemands", "SocialSupport"), p.crit = 0.05, method="D1")

#> Variable excluded at Step 1 is - JobControl

#> Variable excluded at Step 2 is - JobDemands

#> Variable excluded at Step 3 is - Smoking

#> Variable excluded at Step 4 is - SocialSupport

#> Variable excluded at Step 5 is - Gender

  pool_lr$RR_Model

#> $`Step 1`
#>                   est std.err signif   lower   upper     OR   L.OR     U.OR
#> (Intercept)    1.1637  2.2828 0.6102 -3.3104  5.6379 3.2018 0.0365 280.8649
#> Gender        -0.4220  0.4073 0.3001 -1.2202  0.3762 0.6557 0.2952   1.4568
#> Smoking        0.1164  0.3407 0.7327 -0.5515  0.7842 1.1234 0.5761   2.1907
#> Function      -0.1457  0.0421 0.0005 -0.2282 -0.0632 0.8644 0.7959   0.9387
#> JobControl     0.0021  0.0191 0.9145 -0.0355  0.0396 1.0021 0.9652   1.0404
#> JobDemands    -0.0087  0.0357 0.8065 -0.0787  0.0612 0.9913 0.9243   1.0632
#> SocialSupport  0.0249  0.0540 0.6449 -0.0810  0.1308 1.0252 0.9222   1.1397
#> 
#> $`Step 2`
#>                   est std.err signif   lower   upper     OR   L.OR     U.OR
#> (Intercept)    1.3088  1.8407 0.4770 -2.2988  4.9164 3.7019 0.1004 136.5169
#> Gender        -0.4296  0.4012 0.2843 -1.2158  0.3567 0.6508 0.2965   1.4286
#> Smoking        0.1173  0.3406 0.7305 -0.5502  0.7849 1.1245 0.5768   2.1922
#> Function      -0.1450  0.0415 0.0005 -0.2262 -0.0637 0.8651 0.7975   0.9383
#> JobDemands    -0.0089  0.0357 0.8019 -0.0788  0.0610 0.9911 0.9242   1.0628
#> SocialSupport  0.0239  0.0532 0.6532 -0.0804  0.1282 1.0242 0.9228   1.1367
#> 
#> $`Step 3`
#>                   est std.err signif   lower   upper     OR   L.OR    U.OR
#> (Intercept)    0.9989  1.3631 0.4637 -1.6728  3.6705 2.7152 0.1877 39.2713
#> Gender        -0.4128  0.3955 0.2966 -1.1880  0.3624 0.6618 0.3048  1.4367
#> Smoking        0.1177  0.3406 0.7298 -0.5499  0.7852 1.1249 0.5770  2.1929
#> Function      -0.1448  0.0414 0.0005 -0.2260 -0.0636 0.8652 0.7977  0.9384
#> SocialSupport  0.0236  0.0532 0.6575 -0.0806  0.1278 1.0239 0.9225  1.1363
#> 
#> $`Step 4`
#>                   est std.err signif   lower   upper     OR   L.OR    U.OR
#> (Intercept)    1.0465  1.3564 0.4404 -1.6119  3.7050 2.8477 0.1995 40.6483
#> Gender        -0.4113  0.3954 0.2982 -1.1861  0.3636 0.6628 0.3054  1.4385
#> Function      -0.1432  0.0411 0.0005 -0.2237 -0.0627 0.8666 0.7996  0.9393
#> SocialSupport  0.0231  0.0531 0.6644 -0.0811  0.1272 1.0233 0.9221  1.1356
#> 
#> $`Step 5`
#>                 est std.err signif   lower   upper     OR   L.OR    U.OR
#> (Intercept)  1.5889  0.5348 0.0030  0.5407  2.6371 4.8983 1.7172 13.9725
#> Gender      -0.4196  0.3943 0.2872 -1.1924  0.3531 0.6573 0.3035  1.4235
#> Function    -0.1430  0.0411 0.0005 -0.2235 -0.0625 0.8668 0.7997  0.9394
#> 
#> $`Step 6`
#>                 est std.err signif   lower   upper     OR   L.OR   U.OR
#> (Intercept)  1.3143  0.4622 0.0045  0.4084  2.2203 3.7222 1.5043 9.2097
#> Function    -0.1471  0.0408 0.0003 -0.2272 -0.0671 0.8632 0.7968 0.9351
#> 

  pool_lr$multiparm_p

#> $`Step 1`
#>               Chi_sq D1 & RR p-values
#> Gender         1.074           0.3001
#> Smoking        0.117           0.7327
#> Function      11.984           0.0005
#> JobControl     0.012           0.9145
#> JobDemands     0.060           0.8065
#> SocialSupport  0.212           0.6449
#> 
#> $`Step 2`
#>               Chi_sq D1 & RR p-values
#> Gender         1.147           0.2843
#> Smoking        0.119           0.7305
#> Function      12.230           0.0005
#> JobDemands     0.063           0.8019
#> SocialSupport  0.202           0.6532
#> 
#> $`Step 3`
#>               Chi_sq D1 & RR p-values
#> Gender         1.089           0.2966
#> Smoking        0.119           0.7298
#> Function      12.208           0.0005
#> SocialSupport  0.197           0.6575
#> 
#> $`Step 4`
#>               Chi_sq D1 & RR p-values
#> Gender         1.082           0.2982
#> Function      12.147           0.0005
#> SocialSupport  0.188           0.6644
#> 
#> $`Step 5`
#>          Chi_sq D1 & RR p-values
#> Gender    1.133           0.2872
#> Function 12.119           0.0005
#> 
#> $`Step 6`
#>          Chi_sq D1 & RR p-values
#> Function 12.984            3e-04
#>