Pooling and Predictor selection function for Logistic regression models in multiply imputed datasets

psfmi_lr Pooling and backward selection for Logistic regression prediction models in multiply imputed datasets using different selection methods.

psfmi_lr(
  data,
  nimp = 5,
  impvar = NULL,
  Outcome,
  predictors = NULL,
  p.crit = 1,
  cat.predictors = NULL,
  spline.predictors = NULL,
  int.predictors = NULL,
  keep.predictors = NULL,
  knots = NULL,
  method = "RR",
  print.method = FALSE
)

Arguments

data
Data frame with stacked multiple imputed datasets. The original dataset that contains missing values must be excluded from the dataset. The imputed datasets must be distinguished by an imputation variable, specified under impvar, and starting by 1.

nimp
A numerical scalar. Number of imputed datasets. Default is 5.

impvar
A character vector. Name of the variable that distinguishes the imputed datasets.

Outcome
Character vector containing the name of the outcome variable.

predictors
Character vector with the names of the predictor variables. At least one predictor variable has to be defined.

p.crit
A numerical scalar. P-value selection criterium. A value of 1 provides the pooled model without selection.

cat.predictors
A single string or a vector of strings to define the categorical variables. Default is NULL categorical predictors.

spline.predictors
A single string or a vector of strings to define the (restricted cubic) spline variables. Default is NULL spline predictors. See details.

int.predictors
A single string or a vector of strings with the names of the variables that form an interaction pair, separated by a “:” symbol.

keep.predictors
A single string or a vector of strings including the variables that are forced in the model during predictor selection. Categorical and interaction variables are allowed.

knots
A numerical vector that defines the number of knots for each spline predictor separately.

method
A character vector to indicate the pooling method for p-values to pool the total model or used during predictor selection. This can be "D1", "D2", "D3" or "MPR". See details for more information.

print.method
logical vector. If TRUE full matrix with p-values of all variables according to chosen method (under method) is shown. If FALSE (default) p-value for categorical variables according to method are shown and for continuous and dichotomous predictors Rubin’s Rules are used.

Value

An object of class smodsmi (selected models in multiply imputed datasets) from which the following objects can be extracted: imputed datasets as data, selected pooled model as RR_model, pooled p-values according to pooling method as multiparm, predictors included at each selection step as predictors_in, predictors excluded at each step as predictors_out, and impvar, nimp, Outcome, method, p.crit, predictors, cat.predictors, keep.predictors, int.predictors, spline.predictors, knots, print.method, call, model_type, predictors_final for names of predictors in final selection step, fit.formula is the regression formula of start model and predictors_initial for names of predictors in start model.

Details

The basic pooling procedure to derive pooled coefficients, standard errors, 95 confidence intervals and p-values is Rubin's Rules (RR). Specific procedures are available to derive pooled p-values for categorical (> 2 categories) and spline variables. print.method allows to choose between the pooling methods: “D1” is pooling of the total covariance matrix, ”D2” is pooling of Chi-square values, “D3” is pooling Likelihood ratio statistics (method of Meng and Rubin) and “MPR” is pooling of median p-values (MPR rule). Spline regression coefficients are defined by using the rcs function for restricted cubic splines of the rms package. A minimum number of 3 knots as defined under knots is required.

References

Eekhout I, van de Wiel MA, Heymans MW. Methods for significance testing of categorical covariates in logistic regression models after multiple imputation: power and applicability analysis. BMC Med Res Methodol. 2017;17(1):129.

Enders CK (2010). Applied missing data analysis. New York: The Guilford Press.

Meng X-L, Rubin DB. Performing likelihood ratio tests with multiply-imputed data sets. Biometrika.1992;79:103-11.

van de Wiel MA, Berkhof J, van Wieringen WN. Testing the prediction error difference between 2 predictors. Biostatistics. 2009;10:550-60.

Marshall A, Altman DG, Holder RL, Royston P. Combining estimates of interest in prognostic modelling studies after multiple imputation: current practice and guidelines. BMC Med Res Methodol. 2009;9:57.

Van Buuren S. (2018). Flexible Imputation of Missing Data. 2nd Edition. Chapman & Hall/CRC Interdisciplinary Statistics. Boca Raton.

EW. Steyerberg (2019). Clinical Prediction MOdels. A Practical Approach to Development, Validation, and Updating (2nd edition). Springer Nature Switzerland AG.

http://missingdatasolutions.rbind.io/

Examples

  pool_lr <- psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
  predictors=c("Gender", "Smoking", "Function", "JobControl",
  "JobDemands", "SocialSupport"), method="D1")

#> 
#> Pooled model correctly estimated
#>            using a p-value of 1

  pool_lr$RR_Model

#> $`Step 1`
#>                   est std.err signif   lower   upper     OR   L.OR     U.OR
#> (Intercept)   -0.0215  2.4949 0.9931 -4.9246  4.8817 0.9788 0.0073 131.8530
#> Gender        -0.3545  0.4181 0.3966 -1.1741  0.4652 0.7016 0.3091   1.5923
#> Smoking        0.0757  0.3408 0.8244 -0.5924  0.7437 1.0786 0.5530   2.1037
#> Function      -0.1419  0.0434 0.0011 -0.2270 -0.0568 0.8677 0.7969   0.9448
#> JobControl     0.0069  0.0205 0.7370 -0.0335  0.0473 1.0069 0.9670   1.0485
#> JobDemands     0.0023  0.0387 0.9532 -0.0738  0.0784 1.0023 0.9288   1.0815
#> SocialSupport  0.0443  0.0575 0.4408 -0.0685  0.1572 1.0453 0.9338   1.1702
#> 

  pool_lr$multiparm

#> $`Step 1`
#>               Chi_sq p-value D1 & RR
#> Gender         0.719          0.3966
#> Smoking        0.049          0.8244
#> Function      10.698          0.0011
#> JobControl     0.113          0.7370
#> JobDemands     0.003          0.9532
#> SocialSupport  0.594          0.4408
#> 

  pool_lr <- psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
  predictors=c("Gender", "Smoking", "Function", "JobControl",
  "JobDemands", "SocialSupport"), p.crit = 0.05, method="D1")

#> Variable excluded at Step 1 is - JobDemands

#> Variable excluded at Step 2 is - Smoking

#> Variable excluded at Step 3 is - JobControl

#> Variable excluded at Step 4 is - SocialSupport

#> Variable excluded at Step 5 is - Gender

  pool_lr$RR_Model

#> $`Step 1`
#>                   est std.err signif   lower   upper     OR   L.OR     U.OR
#> (Intercept)   -0.0215  2.4949 0.9931 -4.9246  4.8817 0.9788 0.0073 131.8530
#> Gender        -0.3545  0.4181 0.3966 -1.1741  0.4652 0.7016 0.3091   1.5923
#> Smoking        0.0757  0.3408 0.8244 -0.5924  0.7437 1.0786 0.5530   2.1037
#> Function      -0.1419  0.0434 0.0011 -0.2270 -0.0568 0.8677 0.7969   0.9448
#> JobControl     0.0069  0.0205 0.7370 -0.0335  0.0473 1.0069 0.9670   1.0485
#> JobDemands     0.0023  0.0387 0.9532 -0.0738  0.0784 1.0023 0.9288   1.0815
#> SocialSupport  0.0443  0.0575 0.4408 -0.0685  0.1572 1.0453 0.9338   1.1702
#> 
#> $`Step 2`
#>                   est std.err signif   lower   upper     OR   L.OR    U.OR
#> (Intercept)    0.0664  1.9997 0.9735 -3.8606  3.9935 1.0687 0.0211 54.2427
#> Gender        -0.3568  0.4099 0.3841 -1.1602  0.4467 0.6999 0.3134  1.5631
#> Smoking        0.0746  0.3407 0.8266 -0.5931  0.7424 1.0775 0.5526  2.1009
#> Function      -0.1420  0.0434 0.0011 -0.2271 -0.0569 0.8676 0.7969  0.9447
#> JobControl     0.0068  0.0205 0.7397 -0.0336  0.0472 1.0068 0.9670  1.0483
#> SocialSupport  0.0442  0.0574 0.4410 -0.0684  0.1569 1.0452 0.9339  1.1699
#> 
#> $`Step 3`
#>                   est std.err signif   lower   upper     OR   L.OR    U.OR
#> (Intercept)    0.0933  1.9957 0.9627 -3.8259  4.0125 1.0978 0.0218 55.2826
#> Gender        -0.3554  0.4098 0.3858 -1.1587  0.4478 0.7009 0.3139  1.5649
#> Function      -0.1412  0.0431 0.0011 -0.2257 -0.0567 0.8683 0.7980  0.9449
#> JobControl     0.0069  0.0205 0.7358 -0.0334  0.0473 1.0069 0.9671  1.0484
#> SocialSupport  0.0439  0.0574 0.4442 -0.0686  0.1565 1.0449 0.9337  1.1694
#> 
#> $`Step 4`
#>                   est std.err signif   lower   upper     OR   L.OR    U.OR
#> (Intercept)    0.5793  1.3846 0.6757 -2.1359  3.2945 1.7848 0.1181 26.9638
#> Gender        -0.3903  0.3991 0.3281 -1.1726  0.3920 0.6769 0.3096  1.4800
#> Function      -0.1378  0.0423 0.0012 -0.2209 -0.0547 0.8713 0.8018  0.9467
#> SocialSupport  0.0398  0.0561 0.4781 -0.0702  0.1498 1.0406 0.9322  1.1616
#> 
#> $`Step 5`
#>                 est std.err signif   lower   upper     OR   L.OR    U.OR
#> (Intercept)  1.4947  0.5340 0.0051  0.4480  2.5414 4.4579 1.5651 12.6974
#> Gender      -0.4112  0.3965 0.2997 -1.1884  0.3660 0.6629 0.3047  1.4420
#> Function    -0.1354  0.0421 0.0013 -0.2181 -0.0528 0.8733 0.8041  0.9486
#> 
#> $`Step 6`
#>                 est std.err signif   lower   upper     OR   L.OR   U.OR
#> (Intercept)  1.2290  0.4696  9e-03  0.3079  2.1501 3.4178 1.3605 8.5858
#> Function    -0.1399  0.0420  9e-04 -0.2222 -0.0575 0.8695 0.8007 0.9441
#> 

  pool_lr$multiparm

#> $`Step 1`
#>               Chi_sq p-value D1 & RR
#> Gender         0.719          0.3966
#> Smoking        0.049          0.8244
#> Function      10.698          0.0011
#> JobControl     0.113          0.7370
#> JobDemands     0.003          0.9532
#> SocialSupport  0.594          0.4408
#> 
#> $`Step 2`
#>               Chi_sq p-value D1 & RR
#> Gender         0.758          0.3841
#> Smoking        0.048          0.8266
#> Function      10.706          0.0011
#> JobControl     0.111          0.7397
#> SocialSupport  0.594          0.4410
#> 
#> $`Step 3`
#>               Chi_sq p-value D1 & RR
#> Gender         0.752          0.3858
#> Function      10.737          0.0011
#> JobControl     0.114          0.7358
#> SocialSupport  0.586          0.4442
#> 
#> $`Step 4`
#>               Chi_sq p-value D1 & RR
#> Gender         0.956          0.3281
#> Function      10.591          0.0012
#> SocialSupport  0.503          0.4781
#> 
#> $`Step 5`
#>          Chi_sq p-value D1 & RR
#> Gender    1.075          0.2997
#> Function 10.338          0.0013
#> 
#> $`Step 6`
#>          Chi_sq p-value D1 & RR
#> Function 11.115           9e-04
#> 

  pool_lr$predictors_in

#>        Gender Smoking Function JobControl JobDemands SocialSupport
#> Step 1      1       1        1          1          1             1
#> Step 2      1       1        1          1          0             1
#> Step 3      1       0        1          1          0             1
#> Step 4      1       0        1          0          0             1
#> Step 5      1       0        1          0          0             0
#> Step 6      0       0        1          0          0             0