Evaluate model performance of logistic prediction models in Multiply Imputed datasets

psfmi_perform Evaluate Performance of logistic regression models selected with the psfmi_lr function of the psfmi package by using cross-validation or bootstrapping.

psfmi_perform(
  pobj,
  val_method = NULL,
  data_orig = NULL,
  int_val = TRUE,
  nboot = 10,
  folds = 3,
  nimp_cv = 5,
  nimp_mice = 5,
  p.crit = 1,
  BW = FALSE,
  direction = NULL,
  anova_test = "LRT",
  cv_naive_appt = FALSE,
  cal.plot = FALSE,
  plot.indiv = FALSE,
  groups_cal = 10,
  miceImp,
  ...
)

Arguments

pobj
An object of class pmods (pooled models), produced by a previous call to psfmi_lr.

val_method
Method for internal validation. MI_boot for first Multiple Imputation and than bootstrapping in each imputed dataset and boot_MI for first bootstrapping and than multiple imputation in each bootstrap sample, and cv_MI, cv_MI_RR and MI_cv_naive for the combinations of cross-validation and multiple imputation. To use cv_MI, cv_MI_RR and boot_MI, data_orig has to be specified. See details for more information.

data_orig
dataframe of original dataset that contains missing data for methods cv_MI, cv_MI_RR and boot_MI.

int_val
If TRUE internal validation is conducted using bootstrapping or cross-validation. Default is TRUE. If FALSE only apparent performance measures are calculated.

nboot
The number of bootstrap resamples, default is 10. Used for methods boot_MI and MI_boot.

folds
The number of folds, default is 3. Used for methods cv_MI, cv_MI_RR and MI_cv_naive.

nimp_cv
Numerical scalar. Number of (multiple) imputation runs for method cv_MI.

nimp_mice
Numerical scalar. Number of imputed datasets for method cv_MI_RR and boot_MI. When not defined, the number of multiply imputed datasets is used of the previous call to the function psfmi_lr.

p.crit
A numerical scalar. P-value selection criterium used for backward or forward selection during validation. When set at 1, pooling and internal validation is done without backward selection.

BW
Only used for methods cv_MI, cv_MI_RR and MI_cv_naive. If TRUE backward selection is conducted within cross-validation. Default is FALSE.

direction
Can be used together with val_methods boot_MI and MI_boot. The direction of predictor selection, "BW" is for backward selection and "FW" for forward selection.

anova_test
Test statistic used for backward selection with method cv_MI and MI_cv_naive. Default is method "LRT" for the likelihood ratio test. Method "Chisq" is also possible.

cv_naive_appt
Can be used in combination with val_method MI_cv_naive. Default is TRUE for showing the cross-validation apparent (train) and test results. Set to FALSE to only give test results.

cal.plot
If TRUE a calibration plot is generated. Default is FALSE. Can be used in combination with int_val = FALSE.

plot.indiv
If TRUE calibration plots for each separate imputed dataset are generated, otherwise all calibration plots are plotted in one figure.

groups_cal
A numerical scalar. Number of groups used on the calibration plot. Default is 10. If the range of predicted probabilities is too low, less groups can be chosen.

miceImp
Wrapper function around the mice function.

...
Arguments as predictorMatrix, seed, maxit, etc that can be adjusted for the mice function. To be used in combination with validation methods cv_MI, cv_MI_RR and MI_boot. For method cv_MI the number of imputed datasets is fixed at 1 and cannot be changed.

Value

A psfmi_perform object from which the following objects can be extracted: res_boot, result of pooled performance (in multiply imputed datasets) at each bootstrap step of ROC app (pooled ROC), ROC test (pooled ROC after bootstrap model is applied in original multiply imputed datasets), same for R2 app (Nagelkerke's R2), R2 test, Scaled Brier app and Scaled Brier test. Information is also provided about testing the Calibration slope at each bootstrap step as interc test and Slope test. The performance measures are pooled by a call to the function pool_performance. Another object that can be extracted is intval, with information of the AUC, R2, Scaled Brier score and Calibration slope averaged over the bootstrap samples, in terms of: Orig (original datasets), Apparent (models applied in bootstrap samples), Test (bootstrap models are applied in original datasets), Optimism (difference between apparent and test) and Corrected (original corrected for optimism).

Details

For internal validation five methods can be used, cv_MI, cv_MI_RR, MI_cv_naive, MI_boot and boot_MI. Method cv_MI uses imputation within each cross-validation fold definition. By repeating this in several imputation runs, multiply imputed datasets are generated. Method cv_MI_RR uses multiple imputation within the cross-validation definition. MI_cv_naive, applies cross-validation within each imputed dataset. MI_boot draws for each bootstrap step the same cases in all imputed datasets. With boot_MI first bootstrap samples are drawn from the original dataset with missing values and than multiple imputation is applied. For multiple imputation the mice function from the mice package is used. It is recommended to use a minumum of 100 imputation runs for method cv_MI or 100 bootstrap samples for method boot_MI or MI_boot. Methods cv_MI, cv_MI_RR and MI_cv_naive can be combined with backward selection during cross-validation and with methods boot_MI and MI_boot, backward and forward selection can be used. For methods cv_MI and cv_MI_RR the outcome in the original dataset has to be complete.

Vignettes

MI and Cross-validation - Method cv_MI

MI and Cross-validation - Method cv_MI_RR

MI and Cross-validation - Method MI_cv_naive

MI and Bootstrapping - Method boot_MI

MI and Bootstrapping - Method MI_boot

References

Heymans MW, van Buuren S, Knol DL, van Mechelen W, de Vet HC. Variable selection under multiple imputation using the bootstrap in a prognostic study. BMC Med Res Methodol. 2007(13);7:33.

F. Harrell. Regression Modeling Strategies. With Applications to Linear Models, Logistic and Ordinal Regression, and Survival Analysis (2nd edition). Springer, New York, NY, 2015.

Van Buuren S. (2018). Flexible Imputation of Missing Data. 2nd Edition. Chapman & Hall/CRC Interdisciplinary Statistics. Boca Raton.

Harel, O. (2009). The estimation of R2 and adjusted R2 in incomplete data sets using multiple imputation. Journal of Applied Statistics, 36(10), 1109-1118.

Musoro JZ, Zwinderman AH, Puhan MA, ter Riet G, Geskus RB. Validation of prediction models based on lasso regression with multiply imputed data. BMC Med Res Methodol. 2014;14:116.

Wahl S, Boulesteix AL, Zierer A, Thorand B, van de Wiel MA. Assessment of predictive performance in incomplete data by combining internal validation and multiple imputation. BMC Med Res Methodol. 2016;16(1):144.

EW. Steyerberg (2019). Clinical Prediction MOdels. A Practical Approach to Development, Validation, and Updating (2nd edition). Springer Nature Switzerland AG.

http://missingdatasolutions.rbind.io/

Author

Martijn Heymans, 2020

Examples

pool_lr <- psfmi_lr(data=lbpmilr, formula = Chronic ~ Pain + JobDemands + rcs(Tampascale, 3) +
           factor(Satisfaction) + Smoking, p.crit = 1, direction="FW",
           nimp=5, impvar="Impnr", method="D1")
           
pool_lr$RR_model

#> $`Step 1 - no variables removed -`
#>                            term      estimate  std.error  statistic        df
#> 1                   (Intercept) -18.194984918 7.46386123 -2.4377443  66.86346
#> 2                          Pain   0.787297519 0.28791031  2.7345236  23.89602
#> 3                    JobDemands  -0.009136444 0.05316641 -0.1718462  50.22157
#> 4                       Smoking   0.078959504 0.47947339  0.1646796  72.71887
#> 5         factor(Satisfaction)2  -0.439137053 0.58056883 -0.7563910 132.98121
#> 6         factor(Satisfaction)3  -2.255502423 0.71238726 -3.1661184 113.70396
#> 7  rcs(Tampascale, 3)Tampascale   0.404284518 0.19688244  2.0534311  96.31169
#> 8 rcs(Tampascale, 3)Tampascale'  -0.335633226 0.24663776 -1.3608347  28.85941
#>       p.value           OR    lower.EXP  upper.EXP
#> 1 0.017444622 1.253194e-08 5.554584e-15 0.02827388
#> 2 0.011578833 2.197450e+00 1.249812e+00 3.86360901
#> 3 0.864249071 9.909052e-01 8.928447e-01 1.09973554
#> 4 0.869653055 1.082160e+00 4.228202e-01 2.76966768
#> 5 0.450752704 6.445924e-01 2.065832e-01 2.01129312
#> 6 0.001983287 1.048209e-01 2.594487e-02 0.42349084
#> 7 0.042741589 1.498230e+00 1.018565e+00 2.20378006
#> 8 0.184099209 7.148853e-01 4.408532e-01 1.15925410
#> 

res_perf <- psfmi_perform(pool_lr, val_method = "cv_MI", data_orig = lbp_orig, folds=3,
            nimp_cv = 2, p.crit=0.05, BW=TRUE, anova_test = "LRT",
            miceImp = miceImp, printFlag = FALSE)

#> 
#> Imp run 1

#> 
#> fold 1

#> Removed at Step 1 is - JobDemands

#> Removed at Step 2 is - Smoking

#> Removed at Step 3 is - rcs(Tampascale,3)

#> 
#> Selection correctly terminated, 
#> No more variables removed from the model

#> 
#> fold 2

#> Removed at Step 1 is - Smoking

#> Removed at Step 2 is - JobDemands

#> Removed at Step 3 is - rcs(Tampascale,3)

#> 
#> Selection correctly terminated, 
#> No more variables removed from the model

#> 
#> fold 3

#> Removed at Step 1 is - Smoking

#> Removed at Step 2 is - JobDemands

#> 
#> Selection correctly terminated, 
#> No more variables removed from the model

#> 
#> Imp run 2

#> 
#> fold 1

#> Removed at Step 1 is - Smoking

#> Removed at Step 2 is - JobDemands

#> 
#> Selection correctly terminated, 
#> No more variables removed from the model

#> 
#> fold 2

#> Removed at Step 1 is - Smoking

#> Removed at Step 2 is - JobDemands

#> 
#> Selection correctly terminated, 
#> No more variables removed from the model

#> 
#> fold 3

#> Removed at Step 1 is - Smoking

#> Removed at Step 2 is - JobDemands

#> Removed at Step 3 is - rcs(Tampascale,3)

#> 
#> Selection correctly terminated, 
#> No more variables removed from the model

            
res_perf

#> $pool_stats
#>                  Train      Test
#> AUC          0.8935466 0.8367000
#> Scaled Brier 0.4767861 0.3299335
#> R2           0.5814373 0.4682940
#> 
#> $LP_val
#> (Intercept)     lp_test 
#>   0.1811877   0.8586115 
#> 
#> $auc_test
#>             95% Low    AUC 95% Up
#> AUC (logit)  0.7608 0.8367  0.892
#> 

if (FALSE) {
 set.seed(200)
  res_val <- psfmi_perform(pobj, val_method = "boot_MI", data_orig = lbp_orig, nboot = 5,
  p.crit=0.05, BW=TRUE, miceImp = miceImp, nimp_mice = 5, printFlag = FALSE, direction = "FW")
  
  res_val$stats_val
}